+++ tool/grind.dart
@@ -2,7 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:grinder/grinder.dart';
+import 'package:node_preamble/preamble.dart' as preamble;
 import 'package:pub_semver/pub_semver.dart';
 import 'package:yaml/yaml.dart' as yaml;
 
@@ -22,6 +26,41 @@ validate() async {
 
   // Format it.
   Dart.run("bin/format.dart", arguments: ["-w", "."]);
+}
+
+@Task('Publish to npm')
+npm() {
+  var out = 'dist';
+
+  var pubspec = yaml.loadYaml(getFile("pubspec.yaml").readAsStringSync());
+  var homepage = pubspec["homepage"];
+  var fileName = 'index.js';
+
+  // Generate modified dart2js output suitable to run on node.
+  var tempFile = File('${Directory.systemTemp.path}/temp.js');
+
+  Dart2js.compile(File('tool/node_format_service.dart'),
+      outFile: tempFile, categories: 'all');
+  var dart2jsOutput = tempFile.readAsStringSync();
+  File('$out/$fileName').writeAsStringSync('''${preamble.getPreamble()}
+self.exports = exports; // Temporary hack for Dart-JS Interop under node.
+$dart2jsOutput''');
+
+  File('$out/package.json')
+      .writeAsStringSync(const JsonEncoder.withIndent('  ').convert({
+    "name": "dart-style",
+    "version": pubspec["version"],
+    "description": pubspec["description"],
+    "main": fileName,
+    "typings": "dart-style.d.ts",
+    "scripts": {"test": "echo \"Error: no test specified\" && exit 1"},
+    "repository": {"type": "git", "url": "git+$homepage"},
+    "author": pubspec["author"],
+    "license": "BSD",
+    "bugs": {"url": "$homepage/issues"},
+    "homepage": homepage
+  }));
+  run('npm', arguments: ['publish', out]);
 }
 
 /// Gets ready to publish a new version of the package.